+++ src/arch/mod.rs
@@ -2,5 +2,9 @@
 //
 // This is the main mod.rs file for the arch module.
 
+
+
+pub mod riscv;
+
 // Support for x86_64 processors.
 pub mod x64;