+++ src/arch/riscv/mod.rs
@@ -1,0 +1,5 @@
+// src/arch/riscv/mod.rs
+//
+// Handles RISC-V arch.
+
+pub mod rt;