+++ src/data/mod.rs
@@ -2,5 +2,9 @@
 //
 // The data module contains modules for working with data-types, encoding data, decoding data, and using other programming-languages.
 
+
+// JSON
+pub mod json;
+
 // UCS-2
 pub mod ucs2;